implemting and documanting projects api

+++ app/projects/controllers/remove.js
@@ -1,6 +1,6 @@
 const removeService = require('../services/remove');
 const { ok, error } = require('../../../core/util/response');
-
+const accessCheck = require('../services/accessCheck.js');
 /**
  * @api {delete} /api/projects/:id delete
  * @apiName DeleteProject
@@ -11,11 +11,21 @@ const { ok, error } = require('../../../core/util/response');
 
 const remove = async (request, response) => {
     const { id } = request.params;
+    if (!(await accessCheck(request.user, id))) {
+        return error(response, 403, {
+            en: 'you don\'t have access to this project',
+        });
+    }
     const project = await removeService.fetch_project(id);
     if (project === undefined) {
         return error(request, 404, {
             en: 'project not found',
             fa: 'پروژه یافت نشد',
+        });
+    }
+    if (!(await accessCheck(request.user, id))) {
+        return error(response, 403, {
+            en: 'you don\'t have access to this project',
         });
     }
     await removeService.remove_project(id);
